Added Ping command
- Ping command
When a peer doesn't send any command for 30 sec, we send a void "ping" command. 
This command does nothing but keeping the socket alive (and force an exception if it's broken).
This command is rarely sent, but enable logical timeout in connections.py and no more disconnect from valid hosts

- Fixed request queue on TcpServer
Allows for 100 requests in queue, should be plenty.
Added daemon mode on TcpServer threads too.