Peers handler module for Bismuth nodes
Makes peers handling modular.
This class keeps the current code. It's just moving to a class and some plumbing.
Many Globals were converted to class properties.
Will be easier to maintain, to keep thread safe (it's not with the current code)
and to team work on different part of the node.

I tried hard not to modify more than necessary, will be for a next step when this is tested enough.